Implement reading file

+++ dbLib.cpp
@@ -22,50 +22,88 @@
 
 using namespace std;
 
-void    strPrintTime(char* des, time_t& t) {
+void strPrintTime(char *des, time_t &t)
+{
     tm *pTime = gmtime(&t);
     strftime(des, 26, "%Y-%m-%d %H:%M:%S", pTime);
 }
 
-void loadVMDB(char* fName, L1List<VM_Record> &db) {
+void loadVMDB(char *fName, L1List<VM_Record> &db)
+{
     ifstream inFile(fName);
 
-    if (inFile) {
+    //If file does not exist
+    if(inFile.fail()) return;
+
+    if (inFile)
+    {
         string line;
-        getline(inFile , line);// skip the first line
+        getline(inFile, line); // skip the first line
         VM_Record record;
 
-        db.insertHead(record);/// add dummy object
+        db.insertHead(record); /// add dummy object
 
-        while (getline(inFile , line)) {
+        while (getline(inFile, line))
+        {
             /// On Windows, lines on file ends with \r\n. So you have to remove \r
             if (line[line.length() - 1] == '\r')
                 line.erase(line.length() - 1);
-            if (line.length() > 0) {
-                if (parseVMRecord((char*)line.data(), db[0]))/// parse and store data directly
-                    db.insertHead(record);/// add dummy object for next turn
+            if (line.length() > 0)
+            {
+                if (parseVMRecord((char *)line.data(), db[0])) /// parse and store data directly
+                    db.insertHead(record);                     /// add dummy object for next turn
             }
         }
-        db.removeHead();/// remove the first dummy
+        db.removeHead(); /// remove the first dummy
 
         db.reverse();
         inFile.close();
     }
-    else {
+    else
+    {
         cout << "The file is not found!";
     }
 }
 
-bool parseVMRecord(char *pBuf, VM_Record &bInfo) {
+bool parseVMRecord(char *pBuf, VM_Record &bInfo)
+{
     // TODO: write code to parse a record from given line
+    stringstream stream(pBuf);
+    string buf;
+
+    //SKip dont use data
+    getline(stream, buf, ',');
+
+    //Get report time
+    getline(stream, buf, ',');
+    struct tm tm;
+    strptime((char*)buf.data(), "%m/%d/%Y %H:%M:%S", &tm);
+    time_t time_stamp = timegm(&tm);
+    bInfo.timestamp = time_stamp;
+
+    //Get Tag
+    getline(stream, buf, ',');
+    stringstream tag;
+    tag << setfill('0') << setw(4) << buf; //Format string with leading zero
+    tag >> bInfo.id;
+
+    //Get longitude
+    getline(stream, buf, ',');
+    bInfo.longitude = stod(buf);
+
+    //Get latitude
+    getline(stream, buf, ',');
+    bInfo.latitude = stod(buf);
 }
 
-void process(L1List<VM_Request>& requestList, L1List<VM_Record>& rList) {
-    void*   pGData = NULL;
+void process(L1List<VM_Request> &requestList, L1List<VM_Record> &rList)
+{
+    void *pGData = NULL;
     initVMGlobalData(&pGData);
 
-    while (!requestList.isEmpty()) {
-        if(!processRequest(requestList[0], rList, pGData))
+    while (!requestList.isEmpty())
+    {
+        if (!processRequest(requestList[0], rList, pGData))
             cout << requestList[0].code << " is an invalid event\n";
         requestList.removeHead();
     }
@@ -73,17 +111,20 @@ void process(L1List<VM_Request>& requestList, L1List<VM_Record>& rList) {
     releaseVMGlobalData(pGData);
 }
 
-void printVMRecord(VM_Record &b) {
+void printVMRecord(VM_Record &b)
+{
     printf("%s: (%0.5f, %0.5f), %s\n", b.id, b.longitude, b.latitude, ctime(&b.timestamp));
 }
 
 /// This function converts decimal degrees to radians
-inline double deg2rad(double deg) {
+inline double deg2rad(double deg)
+{
     return (deg * __PI / 180);
 }
 
 ///  This function converts radians to decimal degrees
-inline double rad2deg(double rad) {
+inline double rad2deg(double rad)
+{
     return (rad * 180 / __PI);
 }
 
@@ -96,13 +137,14 @@ inline double rad2deg(double rad) {
  * @param lon2d Longitude of the second point in degrees
  * @return The distance between the two points in kilometers
  */
-double distanceEarth(double lat1d, double lon1d, double lat2d, double lon2d) {
+double distanceEarth(double lat1d, double lon1d, double lat2d, double lon2d)
+{
     double lat1r, lon1r, lat2r, lon2r, u, v;
     lat1r = deg2rad(lat1d);
     lon1r = deg2rad(lon1d);
     lat2r = deg2rad(lat2d);
     lon2r = deg2rad(lon2d);
-    u = sin((lat2r - lat1r)/2);
-    v = sin((lon2r - lon1r)/2);
+    u = sin((lat2r - lat1r) / 2);
+    v = sin((lon2r - lon1r) / 2);
     return 2.0 * earthRadiusKm * asin(sqrt(u * u + cos(lat1r) * cos(lat2r) * v * v));
 }
